[Window] The window is now an engine subsystem which is using SDL2. All the complexity has been removed from the Editor. [Input] Mouse, keyboard and controller input is now done using SDL2. [Controller] As a result of using SDL2, PlayStation controllers are now supported. Added some platform specific buttons. [Settings] The output resolution in Spartan.ini, will now define the initial window size (if it exists). Mouse visibility also works. [Shaders] Shader source is now acquired by RHI_Shader and doesn't rely on external include handlers from compiler APIs. [Assert] When in release mode, asserts will now log to file instead of silenting crashing. [Context] Subsystems which failed to initialise will now automatically removed. [PhysicalDevice] Fixed a bug where GPU memory wasn't detected correctly. [Vulkan] Updated to latest. [Editor] Mouse picking will now work regardless of the window's size and position. [Editor] Fixed an ImGui assert related to the progress bar. [Misc] Big code cleanup. [Version] Increased version to 0.33

+++ Runtime/RHI/D3D11/D3D11_Utility.h
@@ -151,7 +151,7 @@ namespace Spartan::d3d11_utility
                 adapter_desc.VendorId,                                      // vendor id
                 RHI_PhysicalDevice_Unknown,                                 // type
                 &name[0],                                                   // name
-                static_cast<uint32_t>(adapter_desc.DedicatedVideoMemory),   // memory
+                static_cast<uint64_t>(adapter_desc.DedicatedVideoMemory),   // memory
                 static_cast<void*>(display_adapter))                        // data
             );
         }
